Deprecate yield in favor of Procedure
