+++ .changeset/violet-papers-smell.md
@@ -1,0 +1,5 @@
+---
+"@neo4j/cypher-builder": patch
+---
+
+Deprecate `Cypher.Yield` type in favour of `Cypher.Procedure`
